% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign.R
\name{get.estimable.coefficients}
\alias{get.estimable.coefficients}
\title{A function to determine which coefficients can be estimated by a partial factorial design.}
\usage{
get.estimable.coefficients(full.design, frac.design)
}
\arguments{
\item{full.design}{(required)}

\item{frac.design}{(required)}
}
\description{
A function to determine which coefficients can be estimated by a partial factorial design.
}
\examples{
library(AlgDesign)
my.exp.design <- list( Shape=c("Rectangle", "Square"), Color=c("Blue", "Red", "Yellow"), DiscountPct=c("10", "20", "30"))
my.full.design <- expand.grid(my.exp.design) #gen.factorial(my.exp.design, factors="all")
my.frac.design <- optFederov(data=my.full.design, nTrials=6)  # 6 is minumum so main effects only
get.estimable.coefficients(my.full.design, my.frac.design)
}
\keyword{doe,}
\keyword{multivariate}
\keyword{testing}
