% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{plot.varimp.logistic}
\alias{plot.varimp.logistic}
\title{A function to create a plot after varimp.logistic()}
\usage{
\method{plot}{varimp.logistic}(.data)
}
\arguments{
\item{.data}{a dataframe created by varimp.logistic() (can be piped)}
}
\value{
a plot and the unmodified dataframe created by varimp.logistic()
}
\description{
This function takes the dataframe created by varimp.logistic() and returns a
plot. It will also pass on the dataset, so it can be used in the middle or
the end of a pipe.
}
\examples{
varimp.logistic(modelFit) \%>\% filter(var_imp > 1) \%>\% varimp.logistic()
}
\keyword{importance}
\keyword{plot}
\keyword{variable}
