% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{plotimp.glm}
\alias{plotimp.glm}
\title{A function to create a plot after imp.glm()}
\usage{
plotimp.glm(.data)
}
\arguments{
\item{.data}{a dataframe created by imp.glm() (can be piped)}
}
\value{
a plot and the unmodified dataframe created by imp.glm()
}
\description{
This function takes the dataframe created by imp.glm() and returns a
plot. It will also pass on the dataset, so it can be used in the middle or
the end of a pipe.
}
\examples{
imp.glm(modelFit) \%>\% filter(p_value<0.05) \%>\% plotimp.glm()
}
\keyword{importance}
\keyword{plot}
\keyword{variable}
