% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift.R
\name{QiniCurve}
\alias{QiniCurve}
\title{A function to plot a Qini curve for two models.}
\usage{
QiniCurve(..., modelnames = NULL)
}
\arguments{
\item{modelnames}{= c("model1", "model2") (optional)}

\item{table1, }{table2, (exactly)}
}
\description{
This function allows you to compare model performace by comparing the Qini curves of two models.
}
\examples{
QiniCurve2(PerfTable_uplift, PerfTable_propensity, modelnames = c("Logit Uplift", "Logit Propensity"))
}
