% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift_plots.R
\name{QiniCurve2}
\alias{QiniCurve2}
\title{A function to plot a Qini curve for two models.}
\usage{
QiniCurve2(table1, table2, modelnames = c("model1", "model2"))
}
\arguments{
\item{table1, }{table2, (exactly)}

\item{modelnames}{= c("model1", "model2") (optional)}
}
\description{
This function allows you to compare model performace by comparing the Qini curves of two models.
}
\examples{
QiniCurve2(PerfTable_uplift, PerfTable_propensity, modelnames = c("Logit Uplift", "Logit Propensity"))
}
