% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uplift.R
\name{QiniTable}
\alias{QiniTable}
\title{A function to compute the performance of an uplift estimator.}
\usage{
QiniTable(data, treat, outcome, prediction, nb.group = 10)
}
\arguments{
\item{data:}{a data frame containing the treatment, the outcome and the predictors.}

\item{treat:}{name of a binary (numeric) vector representing the treatment assignment (coded as 0/1)}

\item{outcome:}{name of a binary response (numeric) vector (coded as 0/1)}

\item{prediction:}{a score to sort the observations from highest to lowest uplift or propensity.}

\item{nb.group:}{(optional, default = 10)}
}
\description{
Adapted from QiniTable in the tools4uplift package
by Florian Zettelmeyer to ensure that the n-tiles are calculated
based on the treatment observations, not the stacked observations..
}
\examples{
QiniTable(
 expdata_stacked.test,
 treat = "ad",
 outcome = "converted",
 prediction = "score_logitFit_treat",
 nb.group = 20
)
}
