% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign.R
\name{ss.2proportion}
\alias{ss.2proportion}
\title{A function to determine the sample size required to estimate the difference of two proportions given a desired confidence level and interval half width}
\usage{
ss.2proportion(
  halfwidth,
  confidence = 0.95,
  proportion1 = 0.5,
  proportion2 = 0.5
)
}
\arguments{
\item{halfwidth}{(required)}

\item{confidence}{(optional with default of .95)}

\item{proportion1}{(optional with default of .5)}

\item{proportion2}{(optional with default of .5)}
}
\description{
A function to determine the sample size required to estimate the difference of two proportions given a desired confidence level and interval half width
}
\examples{
ss.2proportion(0.03)
ss.2proportion(0.03, confidence=0.99)
ss.2proportion(0.03, confidence=0.99, proportion1=.1, proportion1=.2)
}
\keyword{calcuation}
\keyword{sample}
\keyword{size}
