% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign.R
\name{ss.proportion}
\alias{ss.proportion}
\title{A function to determine the sample size required to estimate a proportion given a desired confidence level and interval half width}
\usage{
ss.proportion(halfwidth, confidence = 0.95, proportion = 0.5)
}
\arguments{
\item{halfwidth}{(required)}

\item{confidence}{(optional with default of .95)}

\item{proportion}{(optional with default of .5)}
}
\description{
A function to determine the sample size required to estimate a proportion given a desired confidence level and interval half width
}
\examples{
ss.proportion(0.03)
ss.proportion(0.03, confidence=0.99)
ss.proportion(0.03, confidence=0.99, proportion=.1)
}
\keyword{calcuation}
\keyword{sample}
\keyword{size}
