% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_mktg482}
\alias{update_mktg482}
\title{A function to update the mkgt482 package}
\usage{
update_mktg482()
}
\description{
This function allows you to easily update the mkgt482 package from github
}
\examples{
update_mktg482()
}
\keyword{update}
