% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{varimp.logistic}
\alias{varimp.logistic}
\alias{varimp.logistic.glm}
\alias{varimp.logistic.train}
\title{A function to report variable importance after glm() or either glm or glmnet through caret}
\usage{
varimp.logistic(modelFit)

\method{varimp.logistic}{glm}(modelFit)

\method{varimp.logistic}{train}(modelFit)
}
\arguments{
\item{modelFit}{reguired: estimated/trained glm model}
}
\value{
A tibble with variable, var_imp, p_value, factor, OR_std, OR_sd_perc
}
\description{
This function reports standardized coefficients and ranks variable by importance: The coefficients of continuous variables are standardized to a two standard deviation change of the variable. The coefficients for factor variables are left unchanged. This follows the procedure suggested by Andrew Gelman in "Scaling regression inputs by dividing by two standard deviations," Statistics in Medicine (2008), Vol. 27, pp. 2965-2873.
The function takes as inputs models created by glm or caret using glm
}
\section{Methods (by class)}{
\itemize{
\item \code{glm}: Method for glm()

\item \code{train}: Method for either glm or glmnet in caret. This works regardless of whether the data is pre-processed with "scale"
}}

\examples{
varimp.logistic(logit1)
varimp.logistic(logitFit)
}
\keyword{importance}
\keyword{variable}
