% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainsplot.R
\name{gainsplot}
\alias{gainsplot}
\title{A function to plot a gains curve}
\usage{
gainsplot(..., label.var, bin = 10)
}
\arguments{
\item{bin}{(defaults to 10)}

\item{score1, }{score 2, ... (required)}

\item{label}{(required)}
}
\description{
This function allows you to compare model performace by comparing the gains curves of models.
}
\examples{
gainsplot(logit1$fitted.values, logit2$fitted.values, logit3$fitted.values, label.var = bbb$buyer)
}
\keyword{auc}
\keyword{gains,}
