% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{imp.glm}
\alias{imp.glm}
\alias{imp.glm.glm}
\alias{imp.glm.train}
\title{A function to report variable importance after glm()}
\usage{
imp.glm(modelFit)

\method{imp.glm}{glm}(modelFit)

\method{imp.glm}{train}(modelFit)
}
\arguments{
\item{modelFit}{reguired: estimated/trained glm model}
}
\value{
A tibble with variable, var_imp, p_value, factor, OR_std, OR_sd_perc
}
\description{
This function reports standardized odds ratios and ranks variable by importance: The odds ratios of continuous variables are standardized to a two standard deviation change of the variable. The odds ratios for factor variables are left unchanged. This follows the procedure suggested by Andrew Gelman in "Scaling regression inputs by dividing by two standard deviations," Statistics in Medicine (2008), Vol. 27, pp. 2965-2873.
The function takes as inputs models created by glm or caret using glm
}
\section{Methods (by class)}{
\itemize{
\item \code{glm}: Method for glm()

\item \code{train}: Method for glm in caret. This works regardless of whether the data is pre-processed with "scale"
}}

\examples{
imp.glm(logit1)
imp.glm(logitFit)
}
\keyword{importance}
\keyword{variable}
